Add files created by cookiecutter
The CookieCutter template for Data Science makes a directory. It contains a setup that makes
you projects easier to come back to in the future. With a known folder structuer and a known way to add files to the project is is easier for others to understand what you are doing and,
not least, for you to remember what you did.